pyproject.toml
[project]
name = "stockflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.2",
    "django>=5.2.1",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "faker>=37.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
]
production = [
    "gunicorn>=23.0.0",
]
